Annotated exceptions on public APIs
