+++ App/Program.cs
@@ -11,16 +11,19 @@ using Serilog.Events;
 
 HostApplicationBuilder builder = Host.CreateApplicationBuilder();
 
+// adds Serilog to demonstrate library logging capabilities
 builder.Logging.ClearProviders();
 Log.Logger = new LoggerConfiguration()
-    .MinimumLevel.Debug()
+    .MinimumLevel.Debug() // most verbose stuff is Debug level
     .ReadFrom.Configuration(builder.Configuration)
     .WriteTo.Console(LogEventLevel.Debug)
     .CreateLogger();
 builder.Logging.AddSerilog();
 
+// adds all injectable types as services
 builder.Services.AddHidHide();
 
+// runs example code
 builder.Services.AddHostedService<DemoService>();
 
 IHost app = builder.Build();
